@@ -1,6 +1,6 @@
 {
   "name": "@symphony-ui/uitoolkit",
-  "version": "4.10.3",
+  "version": "5.0.6-experimental",
   "private": true,
   "description": "React components for the Symphony UI design system",
   "license": "ISC",
@@ -17,15 +17,15 @@
     "*.{ts,tsx,js,jsx}": "eslint --cache --fix",
     "*.{js,css,md}": "prettier --write"
   },
-  "dependencies": {},
   "devDependencies": {
     "husky": "^8.0.3"
   },
-  "packageManager": "yarn@1.22.11",
+  "packageManager": "yarn@4.5.1",
   "engines": {
-    "node": ">=14.21.2"
+    "node": ">=22.14.0"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org"
-  }
+  },
+  "stableVersion": "4.10.3"
 }
