@@ -1,13 +1,16 @@
 {
   "name": "@symphony-ui/uitoolkit-components",
-  "version": "4.10.3",
+  "version": "5.0.6-experimental",
   "description": "React components for the Symphony UI design system",
   "publishConfig": {
     "registry": "https://registry.npmjs.org"
   },
   "license": "ISC",
   "sideEffects": false,
-  "main": "index.js",
+  "main": "dist/index.js",
+  "files": [
+    "dist"
+  ],
   "scripts": {
     "prepare": "cd ../.. && husky install",
     "clean": "rm -rf dist",
@@ -29,24 +32,21 @@
     "linkit": "yarn build && cp package.json ./dist && cd ./dist && yarn link && cd .."
   },
   "peerDependencies": {
-    "react": ">=16.8.0",
-    "react-dom": ">=16.8.0"
-  },
-  "engines": {
-    "node": ">=14.21.2"
+    "react": ">=17.0.2",
+    "react-dom": ">=17.0.2"
   },
   "dependencies": {
     "@emotion/cache": "^11.11.0",
     "@emotion/react": "^11.11.4",
     "@floating-ui/react": "^0.26.20",
     "@popperjs/core": "^2.4.4",
-    "@symphony-ui/uitoolkit-styles": "4.10.3",
+    "@symphony-ui/uitoolkit-styles": "5.0.6-experimental",
     "clsx": "^2.1.1",
     "date-fns": "^2.16.1",
     "emoji-js": "^3.5.0",
     "github-markdown-css": "^4.0.0",
     "prop-types": "^15.8.1",
-    "react-resize-detector": "^5.0.3",
+    "react-resize-detector": "^9.1.0",
     "react-select": "^4.3.1",
     "react-virtualized": "^9.22.2",
     "uuid": "^8.3.2"
@@ -77,17 +77,17 @@
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.6.0",
     "@types/enzyme": "^3.10.5",
-    "@types/react": "^16.9.35",
-    "@types/react-dom": "^16.9.8",
+    "@types/react": "^17.0.84",
+    "@types/react-dom": "^17.0.26",
     "@types/react-select": "^4.0.17",
     "@types/react-virtualized": "^9.21.10",
     "@types/styled-components": "5.1.26",
     "@typescript-eslint/eslint-plugin": "^6.9.0",
     "@typescript-eslint/parser": "^6.9.0",
     "@vitest/coverage-v8": "^3.1.1",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.8.0",
     "css-loader": "^4.3.0",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^8.51.0",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-jsx-a11y": "^6.6.1",
@@ -95,15 +95,16 @@
     "eslint-plugin-react-hooks": "^2.5.1",
     "html-react-parser": "^0.14.0",
     "husky": "^8.0.3",
+    "import-meta-resolve": "^4.1.0",
     "lint-staged": "^12.1.2",
     "madge": "^6.1.0",
     "markdown-loader": "^6.0.0",
     "mdx-embed": "^1.1.2",
     "prettier": "^2.2.1",
-    "react": "^16.13.1",
+    "react": "^17.0.2",
     "react-docgen-typescript-loader": "^3.7.2",
     "react-docgen-typescript-plugin": "^1.0.5",
-    "react-dom": "^16.13.1",
+    "react-dom": "^17.0.2",
     "react-hooks-testing-library": "^0.6.0",
     "sass": "^1.26.11",
     "sass-loader": "^10.0.2",
@@ -121,5 +122,6 @@
   "lint-staged": {
     "*.{ts,tsx,js,jsx}": "eslint --cache --fix",
     "*.{css,md}": "prettier --write"
-  }
+  },
+  "stableVersion": "4.10.3"
 }
